Internal: restructure playback controller file
